Simplify login button stories
