+++ client/blocks/login/stories/submit-button/crowdsignal.stories.tsx
@@ -10,7 +10,7 @@ import type { Meta } from '@storybook/react';
 import '../../../../layout/masterbar/crowdsignal.scss';
 
 const meta: Meta = {
-	title: 'client/blocks/Login/Submit Button/Crowdsignal',
+	title: 'client/blocks/Login/Submit Button/Brands',
 	decorators: [ LoginFormAction, LoginFormWrapper, CrowdsignalWrapper ],
 	component: LoginSubmitButton,
 	args: { ...submitButtonArgs },
@@ -18,4 +18,4 @@ const meta: Meta = {
 
 export default meta;
 
-export const Default: SubmitButtonStory = {};
+export const Crowdsignal: SubmitButtonStory = {};
