--- client/blocks/login/stories/submit-button/gravatar.stories.tsx
@@ -1,20 +1,0 @@
-import LoginSubmitButton from '../../login-submit-button';
-import {
-	submitButtonArgs,
-	LoginFormAction,
-	GravatarWrapper,
-	type SubmitButtonStory,
-} from '../shared';
-import type { Meta } from '@storybook/react';
-import '../../../../login/wp-login/style.scss';
-
-const meta: Meta = {
-	title: 'client/blocks/Login/Submit Button/Gravatar',
-	decorators: [ LoginFormAction, GravatarWrapper ],
-	component: LoginSubmitButton,
-	args: { ...submitButtonArgs },
-};
-
-export default meta;
-
-export const Default: SubmitButtonStory = {};
